requirements.txt
pip~=23.2.1
Jinja2~=3.1.5
MarkupSafe~=3.0.2
Werkzeug~=3.1.3
click~=8.1.8
blinker~=1.9.0
itsdangerous~=2.2.0
gunicorn~=23.0.0
packaging~=24.2
Flask~=3.1.0
google-genai
Flask-SocketIO~=5.5.1
eventlet
